Slight restructuring of the Settings view to accomodate Remote Command Error status.
Fix more issues with redundant web service calls.
